Debt - use CSS variables over `registerThemingParticipant` (#165169)
