+++ src/vs/workbench/browser/media/style.css
@@ -3,6 +3,10 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
+/* Animations */
+
+@keyframes fadeIn { from { opacity: 0; } to { opacity: 1; } }
+
 /* Font Families (with CJK support) */
 
 .mac { font-family: -apple-system, BlinkMacSystemFont, sans-serif; }
@@ -45,18 +49,21 @@ body.web {
 	position: fixed; /* prevent bounce effect */
 }
 
-.monaco-workbench.web {
-	touch-action: none; /* Disable browser handling of all panning and zooming gestures. Removes 300ms touch delay. */
-}
-
-@keyframes fadeIn { from { opacity: 0; } to { opacity: 1; } }
-
 .monaco-workbench {
 	font-size: 13px;
 	line-height: 1.4em;
 	position: relative;
 	z-index: 1;
 	overflow: hidden;
+	color: var(--vscode-foreground);
+}
+
+.monaco-workbench.web {
+	touch-action: none; /* Disable browser handling of all panning and zooming gestures. Removes 300ms touch delay. */
+}
+
+.monaco-workbench ::selection {
+	background-color: var(--vscode-selection-background);
 }
 
 .monaco-workbench.border:not(.fullscreen) {
@@ -106,6 +113,15 @@ body.web {
 	font-size: 100%;
 }
 
+.monaco-workbench input::placeholder,
+.monaco-workbench input::-webkit-input-placeholder,
+.monaco-workbench input::-moz-placeholder,
+.monaco-workbench textarea::placeholder,
+.monaco-workbench textarea::-webkit-input-placeholder,
+.monaco-workbench textarea::-moz-placeholder {
+	color: var(--vscode-input-placeholderForeground);
+}
+
 .monaco-workbench select {
 	font-family: inherit;
 }
@@ -135,6 +151,9 @@ body.web {
 	-webkit-app-region: no-drag;
 }
 
+.monaco-workbench .codicon {
+	color: var(--vscode-icon-foreground);
+}
 
 .monaco-workbench .codicon[class*='codicon-'] {
 	font-size: 16px; /* sets font-size for codicons in workbench (https://github.com/microsoft/vscode/issues/98495) */
@@ -147,6 +166,7 @@ body.web {
 /* Custom Dropdown (select) Arrows */
 
 .monaco-workbench select {
+	appearance: none;
 	-webkit-appearance: none;
 	-moz-appearance: none;
 	/* Hides inner border from FF */
@@ -180,12 +200,21 @@ body.web {
 .monaco-workbench select:focus,
 .monaco-workbench input[type="button"]:focus,
 .monaco-workbench input[type="text"]:focus,
+.monaco-workbench button:focus,
 .monaco-workbench textarea:focus,
-.monaco-workbench input[type="checkbox"]:focus {
+.monaco-workbench input[type="search"]:focus,
+.monaco-workbench input[type="checkbox"]:focus ,
+.monaco-workbench .monaco-list:not(.element-focused):focus:before {
 	outline-width: 1px;
 	outline-style: solid;
 	outline-offset: -1px;
+	outline-color: var(--vscode-focusBorder);
 	opacity: 1;
+}
+
+.monaco-workbench.hc-black .synthetic-focus input,
+.monaco-workbench.hc-light .synthetic-focus input {
+	background: transparent; /* Search input focus fix when in high contrast */
 }
 
 .monaco-workbench input[type="checkbox"]:focus {
@@ -219,6 +248,14 @@ body.web {
 	outline-offset: -1px;
 }
 
+.monaco-workbench .monaco-list .monaco-list-row .monaco-highlighted-label .highlight {
+	color: var(--vscode-list-highlightForeground);
+}
+
+.monaco-workbench .monaco-list .monaco-list-row.focused .monaco-highlighted-label .highlight {
+	color: var(--vscode-list-focusHighlightForeground);
+}
+
 .monaco-workbench .synthetic-focus :focus {
 	outline: 0 !important; /* elements within widgets that draw synthetic-focus should never show focus */
 }
@@ -238,4 +275,19 @@ body.web {
 
 .monaco-workbench a.monaco-link:hover {
 	text-decoration: underline; /* render underline on hover for accessibility requirements */
+}
+
+.monaco-workbench .monaco-action-bar:not(.vertical) .action-label:not(.disabled):hover,
+.monaco-workbench .monaco-action-bar:not(.vertical) .monaco-dropdown-with-primary:not(.disabled):hover {
+	background-color: var(--vscode-toolbar-hoverBackground);
+}
+
+.monaco-workbench .monaco-action-bar:not(.vertical) .action-item.active .action-label:not(.disabled),
+.monaco-workbench .monaco-action-bar:not(.vertical) .monaco-dropdown.active .action-label:not(.disabled) {
+	background-color: var(--vscode-toolbar-activeBackground);
+}
+
+.monaco-workbench .monaco-action-bar:not(.vertical) .action-item .action-label:hover:not(.disabled) {
+	outline: 1px dashed var(--vscode-toolbar-hoverOutline);
+	outline-offset: -1px;
 }
