+++ src/vs/workbench/contrib/feedback/browser/media/feedback.css
@@ -52,6 +52,7 @@
 
 .monaco-workbench .feedback-form .content .channels a {
 	padding: 2px 0;
+	color: var(--vscode-textLink-foreground, var(--vscode-contrastBorder));
 }
 
 .monaco-workbench .feedback-form .content .channels a:hover {
@@ -113,6 +114,10 @@
 
 .monaco-workbench .feedback-form .sentiment:hover {
 	background-color: #eaeaea;
+}
+
+.monaco-workbench .feedback-form .sentiment.checked {
+	border: 1px solid var(--vscode-inputOption-activeBorder);
 }
 
 /* Theming */
