+++ src/vs/workbench/contrib/search/browser/searchMessage.ts
@@ -10,7 +10,7 @@ import { parseLinkedText } from 'vs/base/common/linkedText';
 import Severity from 'vs/base/common/severity';
 import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 import { INotificationService } from 'vs/platform/notification/common/notification';
-import { SeverityIcon } from 'vs/platform/severityIcon/common/severityIcon';
+import { SeverityIcon } from 'vs/platform/severityIcon/browser/severityIcon';
 import { TextSearchCompleteMessage, TextSearchCompleteMessageType } from 'vs/workbench/services/search/common/searchExtTypes';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
 import { Schemas } from 'vs/base/common/network';
